requirements.txt
asgiref==3.3.4
astroid==2.5.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2
djangorestframework==3.12.4
httpie==2.4.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pygments==2.8.1
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
